refactor(Modal): Separate Modal into Dialog component and createModal high order component creator
