+++ packages/ringcentral-widgets/components/Dialog/i18n/de-DE.js
