+++ packages/ringcentral-widgets/components/Dialog/i18n/en-GB.js
