+++ packages/ringcentral-widgets/components/Dialog/i18n/es-419.js
