+++ packages/ringcentral-widgets/components/Dialog/i18n/es-ES.js
