+++ packages/ringcentral-widgets/components/Dialog/i18n/fr-CA.js
