+++ packages/ringcentral-widgets/components/Dialog/i18n/fr-FR.js
