+++ packages/ringcentral-widgets/components/Dialog/i18n/index.js
