+++ packages/ringcentral-widgets/components/Dialog/i18n/ja-JP.js
