+++ packages/ringcentral-widgets/components/Dialog/i18n/loadLocale.js
