+++ packages/ringcentral-widgets/components/Dialog/i18n/pt-BR.js
