+++ packages/ringcentral-widgets/components/Dialog/i18n/zh-TW.js
