+++ packages/ringcentral-widgets/components/Dialog/styles.scss
@@ -1,41 +1,11 @@
 @import '../../lib/commonStyles/colors';
 @import '../../lib/commonStyles/fonts';
-.container {
-  position: fixed;
-  top: 20%;
-  left: 0;
+
+.dialog {
+  display: block;
+  position: relative;
   width: 100%;
   height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: baseline;
-  z-index: 10000;
-  @include primary-font;
-}
-
-.containerHidden {
-  position: fixed;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: baseline;
-  z-index: -1;
-}
-
-.modal {
-  min-width: 190px;
-  max-height: 350px;
-  width: 90%;
-  background-color: $snow;
-  box-sizing: border-box;
-  box-shadow: rgba(0, 0, 0, 0.25) 0px 14px 45px, rgba(0, 0, 0, 0.22) 0px 10px 18px;
-  border-radius: 2px;
-  transform: translate3d(0, 0, 0);
-  position: relative; // for defense
-  transition: transform 500ms cubic-bezier(0, 0, 0.25, 1), opacity 500ms cubic-bezier(0, 0, 0.25, 1);
-  z-index: 10002;
   .closeBtn {
     position: absolute;
     top: 8px;
@@ -65,28 +35,6 @@
   }
 }
 
-.modalHidden {
-  transform: translate3d(0, -100px, 0);
-  opacity: 0;
-  z-index: -1;
-  height: 0;
-  width: 0;
-  overflow: hidden;
-}
-
-.mask {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.6);
-  z-index: 10001;
-}
-
-.maskHidden {
-  display: none;
-}
 
 .flatBtn {
   font-size: 12px;
